JS: revamp error notification system
I was originally shooting for retrying AJAX requests, so I built the
system to be able to clear errors after a successful request on the
same key. Thus, the system was out of sync with the present, because at
the moment, no retrying of requests is done. So, I've removed the
unnecessary parts.